+++ tests/test_boundary.py
@@ -30,7 +30,7 @@ class TestCustomFunction:
     def test_call_stages_custom_call_with_python_function(self):
         @af.custom
         def bracket(x):
-            return af.string.format("[{}]", x)
+            return af.string.format("[{x}]", x=x)
 
         ir = af.trace(lambda x: af.string.concat(bracket(x), "!"))("seed")
         custom_eqn = ir.eqns[0]
@@ -51,7 +51,7 @@ class TestCustomFunction:
         def make_custom():
             @af.custom
             def duplicate(x):
-                return af.string.format("[{}]", x)
+                return af.string.format("[{x}]", x=x)
 
             return duplicate
 
@@ -65,7 +65,7 @@ class TestCustomFunction:
     def test_rules_are_stored_in_rule_mappings(self):
         @af.custom
         def bracket(x):
-            return af.string.format("[{}]", x)
+            return af.string.format("[{x}]", x=x)
 
         @bracket.set_pushforward
         def bracket_pushforward(in_tree, /, *, call):
@@ -81,25 +81,25 @@ class TestCustomFunction:
     def test_direct_call_behaves_like_function(self):
         @af.custom
         def bracket(x):
-            return af.string.format("[{}]", x)
+            return af.string.format("[{x}]", x=x)
 
         assert bracket("hello") == "[hello]"
 
     def test_undefined_pushforward_falls_back_to_body_ir(self):
         @af.custom
         def bracket(x):
-            return af.string.format("[{}]", x)
+            return af.string.format("[{x}]", x=x)
 
         ir = af.trace(lambda x: bracket(x))("seed")
         out, tangent = af.pushforward(ir).call(("hello",), ("change",))
 
         assert out == "[hello]"
-        assert tangent == "[change]"
+        assert tangent == "change"
 
     def test_undefined_pullback_falls_back_to_body_ir(self):
         @af.custom
         def bracket(x):
-            return af.string.format("[{}]", x)
+            return af.string.format("[{x}]", x=x)
 
         ir = af.trace(lambda x: bracket(x))("seed")
         out, cotangent = af.pullback(ir).call(("hello",), "feedback")
@@ -110,7 +110,7 @@ class TestCustomFunction:
     def test_undefined_batch_falls_back_to_body_ir(self):
         @af.custom
         def bracket(x):
-            return af.string.format("[{}]", x)
+            return af.string.format("[{x}]", x=x)
 
         ir = af.trace(lambda x: bracket(x))("seed")
         batched = af.batch(ir)
@@ -128,7 +128,7 @@ class TestCustomPushforward:
 
         @af.custom
         def pair_program(x, y):
-            left = af.string.format("left {}", x)
+            left = af.string.format("left {x}", x=x)
             right = af.string.concat(y, "!")
             return left, right
 
@@ -155,13 +155,15 @@ class TestCustomPushforward:
     def test_custom_pushforward_rule_matches_mapping_signature(self):
         @af.custom
         def bracket(x):
-            return af.string.format("[{}]", x)
+            return af.string.format("[{x}]", x=x)
 
         @bracket.set_pushforward
         def bracket_pushforward(in_tree, /, *, call):
             primals, tangents = in_tree
             (dx,) = tangents
-            return call(*primals), af.string.format("custom delta: {}", af.ad.materialize(dx))
+            return call(*primals), af.string.format(
+                "custom delta: {value}", value=af.ad.materialize(dx)
+            )
 
         ir = af.trace(lambda x: af.string.concat(bracket(x), "!"))("seed")
         out, tangent = af.pushforward(ir).call(("hello",), ("small change",))
@@ -173,13 +175,15 @@ class TestCustomPushforward:
     async def test_custom_pushforward_async(self):
         @af.custom
         def bracket(x):
-            return af.string.format("[{}]", x)
+            return af.string.format("[{x}]", x=x)
 
         @bracket.aset_pushforward
         async def bracket_pushforward(in_tree, /, *, call):
             primals, tangents = in_tree
             (dx,) = tangents
-            return call(*primals), af.string.format("async delta: {}", af.ad.materialize(dx))
+            return call(*primals), af.string.format(
+                "async delta: {value}", value=af.ad.materialize(dx)
+            )
 
         ir = af.trace(lambda x: bracket(x))("seed")
         out, tangent = await af.pushforward(ir).acall(("hello",), ("change",))
@@ -190,13 +194,13 @@ class TestCustomPushforward:
     def test_set_pushforward_replaces_default_rule(self):
         @af.custom
         def bracket(x):
-            return af.string.format("[{}]", x)
+            return af.string.format("[{x}]", x=x)
 
         @bracket.set_pushforward
         def bracket_pushforward(in_tree, /, *, call):
             primals, tangents = in_tree
             (dx,) = tangents
-            return call(*primals), af.string.format("push {}", af.ad.materialize(dx))
+            return call(*primals), af.string.format("push {value}", value=af.ad.materialize(dx))
 
         ir = af.trace(lambda x: bracket(x))("seed")
         _, tangent = af.pushforward(ir).call(("hello",), ("change",))
@@ -207,18 +211,20 @@ class TestCustomPushforward:
     async def test_set_pushforward_does_not_replace_async_rule(self):
         @af.custom
         def bracket(x):
-            return af.string.format("[{}]", x)
+            return af.string.format("[{x}]", x=x)
 
         @bracket.set_pushforward
         def bracket_pushforward(in_tree, /, *, call):
             primals, tangents = in_tree
             (dx,) = tangents
-            return call(*primals), af.string.format("sync push {}", af.ad.materialize(dx))
+            return call(*primals), af.string.format(
+                "sync push {value}", value=af.ad.materialize(dx)
+            )
 
         ir = af.trace(lambda x: bracket(x))("seed")
         _, tangent = await af.pushforward(ir).acall(("hello",), ("change",))
 
-        assert tangent == "[change]"
+        assert tangent == "change"
 
 
 class TestCustomPullback:
@@ -231,7 +237,7 @@ class TestCustomPullback:
 
         @af.custom
         def pair_program(x, y):
-            left = af.string.format("left {}", x)
+            left = af.string.format("left {x}", x=x)
             right = af.string.concat(y, "!")
             return left, right
 
@@ -256,14 +262,18 @@ class TestCustomPullback:
     def test_custom_pullback_rule_uses_mlx_argument_order(self):
         @af.custom
         def bracket(x):
-            return af.string.format("[{}]", x)
+            return af.string.format("[{x}]", x=x)
 
         @bracket.set_pullback
         def bracket_pullback(in_tree, /, *, call):
             del call
             (primals, output), cotangent = in_tree
             (x,) = primals
-            return (af.string.format("{} via {} from {}", cotangent, output, x),)
+            return (
+                af.string.format(
+                    "{cotangent} via {output} from {x}", cotangent=cotangent, output=output, x=x
+                ),
+            )
 
         ir = af.trace(lambda x: af.string.concat(bracket(x), "!"))("seed")
         out, cotangent = af.pullback(ir).call(("hello",), "feedback")
@@ -275,14 +285,18 @@ class TestCustomPullback:
     async def test_custom_pullback_async(self):
         @af.custom
         def bracket(x):
-            return af.string.format("[{}]", x)
+            return af.string.format("[{x}]", x=x)
 
         @bracket.aset_pullback
         async def bracket_pullback(in_tree, /, *, call):
             del call
             (primals, output), cotangent = in_tree
             del primals
-            return (af.string.format("async {} via {}", cotangent, output),)
+            return (
+                af.string.format(
+                    "async {cotangent} via {output}", cotangent=cotangent, output=output
+                ),
+            )
 
         ir = af.trace(lambda x: bracket(x))("seed")
         out, cotangent = await af.pullback(ir).acall(("hello",), "feedback")
@@ -293,14 +307,16 @@ class TestCustomPullback:
     def test_set_pullback_replaces_default_rule(self):
         @af.custom
         def bracket(x):
-            return af.string.format("[{}]", x)
+            return af.string.format("[{x}]", x=x)
 
         @bracket.set_pullback
         def bracket_pullback(in_tree, /, *, call):
             del call
             (primals, output), cotangent = in_tree
             del primals
-            return (af.string.format("pull {} {}", output, cotangent),)
+            return (
+                af.string.format("pull {output} {cotangent}", output=output, cotangent=cotangent),
+            )
 
         ir = af.trace(lambda x: bracket(x))("seed")
         _, cotangent = af.pullback(ir).call(("hello",), "feedback")
@@ -311,14 +327,18 @@ class TestCustomPullback:
     async def test_set_pullback_does_not_replace_async_rule(self):
         @af.custom
         def bracket(x):
-            return af.string.format("[{}]", x)
+            return af.string.format("[{x}]", x=x)
 
         @bracket.set_pullback
         def bracket_pullback(in_tree, /, *, call):
             del call
             (primals, output), cotangent = in_tree
             del primals
-            return (af.string.format("sync pull {} {}", output, cotangent),)
+            return (
+                af.string.format(
+                    "sync pull {output} {cotangent}", output=output, cotangent=cotangent
+                ),
+            )
 
         ir = af.trace(lambda x: bracket(x))("seed")
         _, cotangent = await af.pullback(ir).acall(("hello",), "feedback")
@@ -336,7 +356,7 @@ class TestCustomBatch:
 
         @af.custom
         def pair_program(x, y):
-            left = af.string.format("left {}", x)
+            left = af.string.format("left {x}", x=x)
             right = af.string.concat(y, "!")
             return left, right
 
@@ -365,7 +385,7 @@ class TestCustomBatch:
     def test_custom_batch_rule(self):
         @af.custom
         def bracket(x):
-            return af.string.format("[{}]", x)
+            return af.string.format("[{x}]", x=x)
 
         @bracket.set_batch
         def bracket_batch(in_tree, /, *, call):
@@ -375,7 +395,7 @@ class TestCustomBatch:
             (xs,) = values
             (x_axis,) = axes
             assert x_axis is True
-            return [af.string.format("<{}>", x) for x in xs], True
+            return [af.string.format("<{x}>", x=x) for x in xs], True
 
         ir = af.trace(lambda x: bracket(x))("seed")
         batched = af.batch(ir)
@@ -386,7 +406,7 @@ class TestCustomBatch:
     async def test_custom_batch_async(self):
         @af.custom
         def bracket(x):
-            return af.string.format("[{}]", x)
+            return af.string.format("[{x}]", x=x)
 
         @bracket.aset_batch
         async def bracket_batch(in_tree, /, *, call):
@@ -396,7 +416,7 @@ class TestCustomBatch:
             (xs,) = values
             (x_axis,) = axes
             assert x_axis is True
-            return [af.string.format("async <{}>", x) for x in xs], True
+            return [af.string.format("async <{x}>", x=x) for x in xs], True
 
         ir = af.trace(lambda x: bracket(x))("seed")
         batched = af.batch(ir)
@@ -406,7 +426,7 @@ class TestCustomBatch:
     def test_set_batch_replaces_default_rule(self):
         @af.custom
         def bracket(x):
-            return af.string.format("[{}]", x)
+            return af.string.format("[{x}]", x=x)
 
         @bracket.set_batch
         def bracket_batch(in_tree, /, *, call):
@@ -416,7 +436,7 @@ class TestCustomBatch:
             (xs,) = values
             (x_axis,) = axes
             assert x_axis is True
-            return [af.string.format("batch <{}>", x) for x in xs], True
+            return [af.string.format("batch <{x}>", x=x) for x in xs], True
 
         ir = af.trace(lambda x: bracket(x))("seed")
         batched = af.batch(ir)
@@ -427,7 +447,7 @@ class TestCustomBatch:
     async def test_set_batch_does_not_replace_async_rule(self):
         @af.custom
         def bracket(x):
-            return af.string.format("[{}]", x)
+            return af.string.format("[{x}]", x=x)
 
         @bracket.set_batch
         def bracket_batch(in_tree, /, *, call):
@@ -437,7 +457,7 @@ class TestCustomBatch:
             (xs,) = values
             (x_axis,) = axes
             assert x_axis is True
-            return [af.string.format("sync batch <{}>", x) for x in xs], True
+            return [af.string.format("sync batch <{x}>", x=x) for x in xs], True
 
         ir = af.trace(lambda x: bracket(x))("seed")
         batched = af.batch(ir)